requirements.txt
asgiref==3.7.2
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.8
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.5.2
MarkupSafe==3.0.2
msgpack==1.1.0
mysqlclient==2.2.6
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclean==3.0.0
pycparser==2.22
pyOpenSSL==24.3.0
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.4.4
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
zope.interface==7.2
